requirements.txt
glob2==0.5
google-api-core==0.1.4
google-auth==1.4.1
google-auth-oauthlib==0.1.0
google-cloud-bigquery==0.29.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
ijson==2.3
json-rpc==1.10.8
jsonschema==2.6.0
numpy==1.12.1
numpydoc==0.7.0
pandas==0.22.0
pandas-gbq==0.4.1
pathlib==1.0.1
pathlib2==2.3.0
scipy==1.0.0
tqdm==4.19.4
ujson==1.35
